### 2023-06-01 #5514 @editorialbot
[PRE REVIEW]: CRATE: A Python package to perform fast material simulations
**Submitting author:** @BernardoFerreira (<a href="http://orcid.org/0000-0001-5956-3877">Bernardo P. Ferreira</a>)
**Repository:** https://github.com/bessagroup/CRATE
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.2
