### 2023-06-09 #5514 @Kevin-Mattheus-Moerman
@BernardoFerreira  thanks for this submission. I am the AEiC on this track and here to help with initial steps. For the moment, could you see if you can address that potentially missing/invalid DOI :point_up: ? You can edit the bib file and call `@editorialbot check references` to check the DOIs again. Thanks.
