### 2023-06-10 f7bebf4
Update references.bib
Fixed JOSS paper reference DOI.
