+++ docs/joss_paper/references.bib
@@ -42,7 +42,7 @@
   pages = {114726},
   year = {2022},
   issn = {0045-7825},
-  doi = {https://doi.org/10.1016/j.cma.2022.114726},
+  doi = {10.1016/j.cma.2022.114726},
   url = {https://www.sciencedirect.com/science/article/pii/S0045782522000895},
   author = {Ferreira, B.P., and Andrade Pires, F.M., and Bessa, M.A.},
 }
