### 2023-06-10 627b5e9
Update setup.cfg
Fixed configuration file GitHub repository URL.
