+++ setup.cfg
@@ -7,7 +7,7 @@ description = CRATE: Clustering-based Nonlinear Analysis of Materials
 long_description = file: README.md
 long_description_content_type = text/markdown
 keywords = simulation, clustering, materials, homogenization, modeling
-url = https://github.com/BernardoFerreira/CRATE
+url = https://github.com/bessagroup/CRATE
 license = BSD-3-Clause License
 license_files = file: LICENSE
 classifiers =
