### 2023-06-26 #5594 @editorialbot
[REVIEW]: CRATE: A Python package to perform fast material simulations
**Submitting author:** @BernardoFerreira (<a href="http://orcid.org/0000-0001-5956-3877">Bernardo P. Ferreira</a>)
**Repository:** https://github.com/bessagroup/CRATE
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.3

@RahulSundar & @atzberg & @Extraweich & @kingyin3613, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
