### 2023-06-27 #5594 @BernardoFerreira
Hello @Extraweich,

Thank you for your kind words and for taking the time to review this project!

Concerning the points you raised:

- Statement of Need:
  - Thank you for pointing this out, I completely missed that! I'll add a concise Statement of Need to Docs (both GitHub and Sphinx front pages) as you suggested.

- Automated tests:

  - You are right, at the moment I don't have an automated testing suit in the workflow, but that is definitely a next step. Can you recommend me any sources that I may follow to properly set one in a standard way?
  - In the context of multi-scale simulations of heterogeneous materials, analytical solutions are rarely available (perhaps for very simple microstructures and constitutive linear behavior). Therefore, the validation of reduced-order models is usually done by comparison with the solution obtained with a Direct Numerical Simulation (DNS), e.g., Finite Element Method, which is taken as the 'ground-truth' or 'high-fidelity solution'. In this context, I provide the DNS solution of each benchmark for validation purposes (as explained [here](https://bessagroup.github.io/CRATE/rst_doc_files/validation/benchmarks.html)). Although an automated test suit to assert the different functions and methods is definitely important, these benchmarks are fundamental to assess the whole simulation process and solution;
  - From what I read in the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests), the automated tests are strongly encouraged (GOOD), but properly documented and reproducible benchmarks to assert the behavior seem to render an OK.
