### 2023-06-27 #5594 @BernardoFerreira
@Extraweich, I added the Statement of Need as you requested (you may check it [here](https://github.com/bessagroup/CRATE) and [here](https://bessagroup.github.io/CRATE/#))!
