### 2023-06-30 #5594 @Kevin-Mattheus-Moerman
@BernardoFerreira on automated testing. It is not a strict requirement to have fully automated testing but it is highly encouraged. What is required at a minimum is that a user can run a testing suite (e.g. manually) to compare to known results. In this case though I would encourage you to explore adding automated unit tests for instance.

You mention that comparison to analytical tests may be challenging. However, it should be relatively straightforward for you to add some sort of verification and/or validation-type tests, which are usually “integration” tests because they test the overall behavior of the software, the typical sort of automated tests that are standard in Python packages (and of course other languages) are unit tests, which confirm that individual functions are working properly.

So although I agree a full "test" e.g. of something like an entire CFD or FEA package may be challenging, it should be much easier to confirm that individual functions are working properly given various known inputs.

So, given the above, I would encourage you to include some form of automated functionality testing / unit tests.

Here are some resources our board members suggested when they helped me formulate this response:
https://github.com/jbusecke/cookiecutter-science-project
https://realpython.com/python-testing/
https://caam37830.github.io/book/09_computing/unittest.html
https://coderefinery.github.io/testing/

@Extraweich perhaps you can comment on the nature of the benchmarking comparisons you mentioned. Are these manual evaluations for large functional blocks (combinations) of code? Do you agree that additional and automated testing of the behaviour of functional units, as I allude to above, should still be achievable in this case?
