### 2023-06-30 #5594 @BernardoFerreira
Hello @Extraweich,

Concerning the points you raised:

- Comparison with other packages:

    - Yes, I believe that there are no other open-source software packages to perform clustering-based reduced-order simulations of materials out there and against which I can compare the performance or functionality of CRATE;
    - Thank you for pointing these open-source projects that I didn't know about! In fact, I'm quite familiar with the authors (researchers) work. To the best of my knowledge, these methods are Direct Numerical Simulation (DNS) methods (similar to FEM) based on a FFT-based homogenization approach. Therefore, it doesn't make sense to compare the functionality/performance directly with CRATE. Nonetheless, these methods can be made available in CRATE's interface "DNS Solver", similar to the FFT-based homogenization basic scheme that is already implemented there! I'll add them to the documentation and search for other related packages as well - does this answer your point?

- Contributing to the software:

    - Given this is the first time I'm publishing an open-source software, I'm still not sure what is the most effective way to promote "anonymous" contributions (some contributions are already being prepared from some of my coworkers though!). I will take a look into your repository example and add a CONTRIBUTING.md file to my project (even if simple) as soon as I go through these first contributions - does this answer your point?

- Automated testing (also @Kevin-Mattheus-Moerman):

    - I completely agree with both of you and I know that including automated functionality testing / unit tests is in fact a standard (and recommended) practice in general software projects. Given the limited amount of time that I have at the moment due to other projects, this is something that I'm planning to implement sequentially as I find time to go through the different modules and build a robust unit testing set for each of them (thank you for providing me those sources @Kevin-Mattheus-Moerman, they will definitely be useful!).
    - This being said, I would like to reinforce that most of CRATE classes/methods are not very prone to be "unit tested", given that they operate on fairly complex data structures and are not meant to be used outside of a full simulation. Hence, although I'm surely able to implement some unit testing (previous point) in some parts of the code, I believe that the full validation against DNS simulations is the primary source of the code's trusthworthiness in this particular case. As we keep developing different research projects with this code, the idea is to keep adding more benchmarks with different application cases (e.g., different materials, microstructures, loadings), the corresponding 'reference' results (obtained with DNS methods) and, if available, results from the literature - this provides validation means and reproducibility at the same time!

@Extraweich, once again thank you for taking your time to review this project. Please let me know your thoughts on the points above!
