### 2023-07-04 #5594 @Extraweich
@BernardoFerreira:

- Comparison with other packages:
In my opinion, it would still be good to mention at least a few relevant packages in the paper, even if their scopes are somewhat different. I would say that the scope (homogenization in materials science / engineering) and the mathematics behind it (FFT-based solution of the Lippmann-Schwinger equation) are similar enough to warrant a comparison. When comparing these packages, it would be a good idea to highlight how CRATE is different (clustering-based model reduction) and what this functionality allows the user to do that they could not do with other packages. Would you agree with me on this point, or am I wrong in assuming that the packages have great similarities up to a point?

- Contributing to the software:
I understand that it is a bit difficult to contribute to your software, but I would still give clear instructions to interested users on how to contribute. So if you would add a CONTRIBUTING.md I would be fully satisfied. :)

- Automated testing:
I also see that the functionality of your code is complex, which makes unit testing a bit difficult. Since @Kevin-Mattheus-Moerman stated, that automated testing is not a strict necessity and given the benchmarks in your repository and your PHD thesis, I would check the given point from my list. I would encourage you to add some unit tests in the future, though.

---

Another thought that came to mind and has nothing to do with the publishing process here. I personally am not very familiar with the methods used in CRATE. I have some basic knowledge of FFT-based solutions to the Lippmann-Schwinger equation, since Matti Schneider (whom you cite once or twice in your dissertation) is remotely involved in my own dissertation project. Nevertheless, I found it somewhat tedious to fully grasp the potential of CRATE. That's because it's certainly a complicated topic. You give a lot of well-chosen benchmarks, but they kind of run themselves, and the user might feel a bit overwhelmed by the abundance of results. What do you think about adding a Jupyter notebook with a very simple step-by-step example to give the user the opportunity to more easily understand how she/he can incorporate CRATE's functionality into her/his specific research challenges?
