### 2023-07-05 fe81812
Add CODE_OF_CONDUCT.md and CONTRIBUTING.md.
