+++ CONTRIBUTING.md
@@ -1,0 +1,30 @@
+# Contributing to CRATE
+
+If you wish to contribute to this project, start by reading CRATE's
+[Code of Conduct](https://github.com/bessagroup/CRATE/blob/master/CODE_OF_CONDUCT.md)
+and act accordingly during the whole contribution process.
+
+Any contribution to this project should follow two main steps:
+
+- **Step 1:** [Open a new discussion](https://github.com/bessagroup/CRATE/discussions) on CRATE's GitHub Discussions space
+
+    - Provide a clear description of your idea, including any coding details that you may already have in mind;
+    
+    - Enroll in a discussion with the project's owners about the merits and pros/cons of your idea;
+    
+    - Proceed to Step 2 if and only if owners accept your idea.
+    
+- **Step 2:** Contribute to CRATE by creating a Pull Request
+
+    - Fork [CRATE's GitHub repository](https://github.com/bessagroup/CRATE) to your GitHub account;
+    - Clone the forked repository into a local directory, thus creating your local repository;
+    - Set your remote forked repository as the 'origin' remote (should be set by default);
+    - Set [CRATE's GitHub repository](https://github.com/bessagroup/CRATE) as the 'upstream' remote;
+    - Pull the most recent version of CRATE's 'master' branch into your local repository;
+    - Create and new development branch and commit your contribution;
+    - Push your contribution to the 'origin' remote;
+    - Create a Pull Request from your forked repository ('head repository') to [CRATE's GitHub repository](https://github.com/bessagroup/CRATE) ('base repository');
+    - The Pull Request is reviewed by the project's owners and your contribution is merged depending on the outcome of the review process.
+    
+Please refer to the several sources available on how to contribute to a GitHub repository (Step 2) (e.g., [source 1](https://www.dataschool.io/how-to-contribute-on-github/), [source 2](https://dev.to/codesphere/how-to-start-contributing-to-open-source-projects-on-github-534n)).
+
