+++ docs/joss_paper/paper.md
@@ -29,19 +29,22 @@ bibliography: references.bib
 
 # Summary
 
-CRATE (Clustering-based Nonlinear Analysis of Materials) is a Python project (package `cratepy`) developed in the context of computational mechanics [@ferreira:2022a] to aid the design and development of new materials. Its main purpose is performing multi-scale nonlinear analyses of heterogeneous materials through a suitable coupling between first-order computational homogenization and clustering-based reduced-order modeling. This means that, given a representative volume element of the material micro-structure and the corresponding material phase properties, `cratepy` computes the material's effective mechanical response under a given loading by leveraging a so-called clustering-based reduced-order model (CROM).
+[CRATE](https://github.com/bessagroup/CRATE) (Clustering-based Nonlinear Analysis of Materials) is a Python project (package [`cratepy`](https://pypi.org/project/cratepy/)) developed in the context of computational mechanics [@ferreira:2022a] to aid the design and development of new materials. Its main purpose is performing multi-scale nonlinear analyses of heterogeneous materials through a suitable coupling between first-order computational homogenization and clustering-based reduced-order modeling. This means that, given a representative volume element of the material micro-structure and the corresponding material phase properties, [`cratepy`](https://pypi.org/project/cratepy/) computes the material's effective mechanical response under a given loading by leveraging a so-called clustering-based reduced-order model (CROM).
 
-![Logo of CRATE (`cratepy`). \label{fig:crate_logo_horizontal_long}](crate_logo_horizontal_long.png)
+![Logo of CRATE ([`cratepy`](https://pypi.org/project/cratepy/)). \label{fig:crate_logo_horizontal_long}](crate_logo_horizontal_long.png)
 
 # Statement of need
 
-CRATE (Clustering-based Nonlinear Analysis of Materials) is a Python project (package `cratepy`) developed in the field of computational mechanics and material science. To the best of the authors' knowledge, it is a first-of-its-kind open-source software that allows any material development enthusiast to perform multi-scale analyses of materials by taking advantage of the recent family of clustering-based reduced-order models (CROMs). \autoref{fig:crate_conceptual_scheme} provides a simple illustration of a CRATE simulation. It is worth remarking that CRATE is supported by a rich documentation that provides a conceptual overview of the project, clear installation instructions, a step-by-step basic workflow description, and detailed guidelines for advanced customized developments. Moreover, `cratepy` relies solely on a few well-established third-party Python scientific computing packages, such as `numpy` [@harris2020array] and `scipy` [@2020SciPy-NMeth], `cratepy`'s modules are extensively documented, and the automatically generated API provides a complete and updated description of the underlying object-oriented implementation, including LaTeX rendered formulae to improve comprehension.
[…]
