+++ docs/joss_paper/references.bib
@@ -107,3 +107,15 @@
  pages={2825--2830},
  year={2011}
 }
+
+@article{Ospald2019,
+        author = {F. Ospald},
+        title = {fibergen: An introductory tool for {FFT}-based material homogenization},
+        year  = {2019},
+        publisher = {The Open Journal},
+        journal = {Journal of Open Source Software},
+        volume = {4},
+        number = {34},
+        pages = {1027},
+        doi = {10.21105/joss.01027}
+}
