### 2023-07-05 #5594 @BernardoFerreira
@Extraweich :

Thank you for your comments and suggestions!

- Comparison with other packages:
    - Updated the paper as you suggested, including references to both open-source projects that you mentioned and highlighting the difference with respect to CRATE.

- Contributing to the software:
    - Updated the repository as you suggested, including both CONTRIBUTING.md and CODE_OF_CONDUCT.md files.

- Automated testing:
    - Thank you for your comprehension. Adding some unit tests is definitely on the task list!

----

Regarding your suggestion:

- I'm glad that you mention this because I'm currently preparing a short course (based on jupyter notebooks) with that goal!
- Nonetheless, and given that you may also want to take advantage of CRATE in your research, the goal of the section "BASIC USAGE" on the Docs is to guide you through the diferent steps to build your own simulations. All the benchmark simulations follow exactly that step-by-step process. The "ADVANCED SECTION" further explains how you can customize the actual code by including new features related with your research (e.g., new material constitutive models).
