### 2023-07-12 #5594 @Kevin-Mattheus-Moerman
@atzberg it says ` Email replies do not support Markdown`, if you instead call `@editorialbot generate my checklist` here in a comment you get a proper list.
