### 2023-07-12 #5594 @Kevin-Mattheus-Moerman
@openjournals/dev how do you suggest to solve this "email replies" issue? Do we just repeat the call in a comment here? Seems odd that it doesn't re-interpret the markdown. Should be a GitHub option to say, re-interpret.
