### 2023-07-12 #5594 @xuanxu
Yes, it looks like GitHub won't allow markdown in from-email comments even if edited later here. I guess for now our only option is to delete that comment and then @atzberg needs to repeat the call from GitHub instead of from an email.
