### 2023-07-16 #5594 @kingyin3613
Hi @BernardoFerreira , I went through the paper, code and documentation, lovely package and well-written documentation! Everything went smoothly, except when I was running the benchmark, there was an encoding issue, I attached details in https://github.com/bessagroup/CRATE/issues/10 . I'm looking for your solutions!
