### 2023-07-17 #5594 @atzberg
[minor item]  The CRATE logo image has black text on the homepage README.md that does not display on github when using the "dark theme."   Maybe change the black text to grey or outline in white or use white background, so it is visible in the "dark theme."  Overall, a nicely organized and documented package.
