### 2023-07-17 89f5148
Attempt to fix enconding issue (greek characters) when writing to screen file.
