### 2023-07-17 bessagroup/CRATE#10 @BernardoFerreira
Hello @kingyin3613,

Thank you for reporting this!

It seems that colors and greek letters are not being properly encoded by your OS - Are you running CRATE on Windows?

I believe that I found a similar problem discussed [here](https://stackoverflow.com/questions/27092833/unicodeencodeerror-charmap-codec-cant-encode-characters), where some comments suggest that I must be explicit about the encoding (UTF-8) when running on Windows OS.

I attempted to fix CRATE accordingly, but I am not facing this issue in either Linux or MacOS. Can you please check if the latest version available on CRATE's GitHub repository (master branch) solved the issue?

Thank you!
