### 2023-07-17 #5594 @BernardoFerreira
> Hi @BernardoFerreira , I went through the paper, code and documentation, lovely package and well-written documentation! Everything went smoothly, except when I was running the benchmark, there was an encoding issue, I attached details in [bessagroup/CRATE#10](https://github.com/bessagroup/CRATE/issues/10) . I'm looking for your solutions!

Hello @kingyin3613!

Thank you for your kind words and for taking your time to review this project!

I've replied to your issue and attempted to fix the issue (displaying colors and greek letters on terminal). I hope that this solves the problem and that you can run the benchmarks successfully - let me know if that is not the case!
