### 2023-07-17 #5594 @BernardoFerreira
> I have now updated the checklist. Overall, it is a well-organized and well-documented software package.

Hello @atzberg,

Thank you for your appreciation and for taking your time to review CRATE!
