### 2023-07-17 ca5d31c
Update CRATE logo with white background.
