### 2023-07-17 #5594 @BernardoFerreira
> [minor item] The CRATE logo image has black text on the homepage README.md that does not display on github when using the "dark theme." Maybe change the black text to grey or outline in white or use white background, so it is visible in the "dark theme." Overall, a nicely organized and documented package.

Thank you for pointing this out! I have fixed this problem by adding the logo with the white background (compatible with both light/dark themes).

P.S. I actually have a CRATE logo version for black background as well, but it doesn't seem straightforward to set a different logo to be displayed according to your theme on GitHub

![crate_logo_black](https://github.com/openjournals/joss-reviews/assets/25851824/f423ec13-e448-4769-a1da-0f7740ddcf18)
