### 2023-07-18 bessagroup/CRATE#10 @kingyin3613
Yes, I'm using Windows. The encoding error was totally gone with the latest version of CRATE, I think the functionality of this package on my Windows OS has been verified. Although I can still see "←[xx m" in the console (see below), but it is just a minor display issue, you may want to fix this in future versions. Thanks for the quick updates @BernardoFerreira !

```
←[32m

End phase: ←[37mSolve reduced microscale equilibrium problem (phase duration time = 7.52e+01s)
--------------------------------------------------------------------------------------------←[0m

←[37m~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~←[0m
←[33mEnding program execution at: ←[0m20h42m35s (17/Jul/2023)

←[33mProblem analysed: ←[0mexample_1_uniaxial_tension_nc54

←[33mTotal execution time: ←[0m3.04e+02s (~0h5m)

←[33mExecution times:

←[0m    Phase                                                 Duration (s)      %
    Read input data                                         5.89e-01       0.19
    Compute cluster analysis data matrix                    3.44e+00       1.13
    Perform RVE cluster analysis                            1.38e+01       4.53
    Compute cluster interaction tensors                     9.91e+01      32.56
    Solve reduced microscale equilibrium problem            7.52e+01      24.70
    Accumulated post-processing operations                  1.12e+02      36.81

                                 ←[32mProgram Completed←[0m
```
