### 2023-07-18 #5594 @atzberg
Hi @BernardoFerreira,  Thanks for fixing items.  The package also looks good to accept. I recommend for publication on JOSS.
