### 2023-07-18 #5594 @Extraweich
> > [minor item] The CRATE logo image has black text on the homepage README.md that does not display on github when using the "dark theme." Maybe change the black text to grey or outline in white or use white background, so it is visible in the "dark theme." Overall, a nicely organized and documented package.
>
> Thank you for pointing this out! I have fixed this problem by adding the logo with the white background (compatible with both light/dark themes).
>
> P.S. I actually have a CRATE logo version for black background as well, but it doesn't seem straightforward to set a different logo to be displayed according to your theme on GitHub
>
> ![crate_logo_black](https://user-images.githubusercontent.com/25851824/254048754-f423ec13-e448-4769-a1da-0f7740ddcf18.png)

Hi @BernardoFerreira, in case you have not found this functionality and are interested in using it, Markdown on GitHub can do this. Check out the raw [README.md](https://raw.githubusercontent.com/Extraweich/homopy/main/README.md) in one of my repositories (I would post the code snippet here, but then it will be rendered, which is not my intention...). You will have to be careful when uploading such a README.md to pages like pypi.org, though, because general Markdown does not support this feature and you'll see some artefacts (see this [example](https://pypi.org/project/homopy/1.0.9/)).
