### 2023-07-18 #5594 @BernardoFerreira
> > > Hi @BernardoFerreira , I went through the paper, code and documentation, lovely package and well-written documentation! Everything went smoothly, except when I was running the benchmark, there was an encoding issue, I attached details in [bessagroup/CRATE#10](https://github.com/bessagroup/CRATE/issues/10) . I'm looking for your solutions!
> >
> >
> > Hello @kingyin3613!
> > Thank you for your kind words and for taking your time to review this project!
> > I've replied to your issue and attempted to fix the issue (displaying colors and greek letters on terminal). I hope that this solves the problem and that you can run the benchmarks successfully - let me know if that is not the case!
>
> Hi @BernardoFerreira , I could successfully run all the benchmarks with the latest version of CRATE, although there was still a minor display issue on Windows OS (or probably only on my Windows system), you may want to fix it in the next few versions. Overall, CRATE has already met all my requirements as a mature and well-documented open-source software, I recommend this package, for publication on JOSS. Thanks for your quick responses to my issues and your contribution to the open-source software community!

Thank you once again @kingyin3613!
