### 2023-07-25 #5594 @RahulSundar
I have updated all the items in the checklist. I went through the paper, code and the documentation. I didn't face any issues while running the benchmark cases and the documentation at the current state is done well @BernardoFerreira.
