### 2023-07-31 #5594 @BernardoFerreira
@Kevin-Mattheus-Moerman:
- Released new CRATE version (1.0.3) after the JOSS review process - this is the version to be used in the paper;
