### 2023-07-31 #5594 @BernardoFerreira
> @BernardoFerreira the version tag on your repository is actually `v1.0.3`, not `1.0.3`. Could you update the Zenodo archive to have the `v` as well? We need to use the official GitHub repository tag. Thanks

Sorry about that, fixed it!
