### 2023-07-31 #5594 @BernardoFerreira
> 👋 @openjournals/bcm-eics, this paper is ready to be accepted and published.
>
> Check final proof point_rightpage_facing_up [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05594/joss.05594/10.21105.joss.05594.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#4440](https://github.com/openjournals/joss-papers/pull/4440), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

@Kevin-Mattheus-Moerman, checked the final proof, didn't find any issues!
