### 2023-07-31 d92e3f0
Added CITATION.cff file after JOSS publication.
