### 2023-07-31 805e5d8
Update README.md
Included JOSS paper citation.
