+++ README.md
@@ -19,25 +19,47 @@
 
 In the particular case of a **research environment**, [cratepy](https://pypi.org/project/cratepy/) is designed to easily accommodate further developments, either by improving the already implemented methods or by including new numerical models and techniques. It also provides all the fundamental means to perform comparisons with alternative methods, both in terms of accuracy and computational cost. In a **teaching environment**, [cratepy](https://pypi.org/project/cratepy/) is a readily available tool for demonstrative purposes and/or academic work proposals in solid mechanics and material-related courses.
 
+Consider leaving a star if you think CRATE is useful for the research community!
+
 ### Authorship & Citation
 CRATE was originally developed by Bernardo P. Ferreira<sup>[1](#f1)</sup> in the context of his PhD Thesis<sup>[2](#f2)</sup>.
-
-If you use CRATE in a scientific publication, it is appreciated that you cite this PhD Thesis:
-
-```
-@phdthesis{ferreira:2022a,
-  title = {Towards Data-driven Multi-scale Optimization of Thermoplastic Blends: Microstructural Generation, Constitutive Development and Clustering-based Reduced-Order Modeling},
-  author = {Ferreira, B.P.},
-  year = {2022},
-  langid = {english},
-  school = {University of Porto},
-  url={https://hdl.handle.net/10216/146900}
-}
-```
 
 <sup id="f1"> 1 </sup> Profile: [LinkedIN](https://www.linkedin.com/in/bpferreira/), [ORCID](https://orcid.org/0000-0001-5956-3877), [ResearchGate](https://www.researchgate.net/profile/Bernardo-Ferreira-11?ev=hdr_xprf)
 
 <sup id="f2"> 2 </sup> Ferreira, B.P. (2022). *Towards Data-driven Multi-scale Optimization of Thermoplastic Blends: Microstructural Generation, Constitutive Development and Clustering-based Reduced-Order Modeling.* PhD Thesis, University of Porto (see [here](http://dx.doi.org/10.13140/RG.2.2.33940.17289))
+
+If you use CRATE in your research or in a scientific publication, it is appreciated that you cite the two papers below.
+
+**Journal of Open Source Software**:
+```
+@article{Ferreira2023,
+  title = {CRATE: A Python package to perform fast material simulations},
+  author = {Bernardo P. Ferreira and F. M. Andrade Pires and Miguel A. Bessa}
+  doi = {10.21105/joss.05594},
+  url = {https://doi.org/10.21105/joss.05594},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {87},
+  pages = {5594},
+  journal = {Journal of Open Source Software}
+}
+```
+
+**Computer Methods in Applied Mechanics and Engineering**:
+```
+@article{Ferreira2022,
+  title = {Adaptivity for clustering-based reduced-order modeling of localized history-dependent phenomena},
+  author = {Ferreira, B.P., and Andrade Pires, F.M., and Bessa, M.A.},
+  doi = {10.1016/j.cma.2022.114726},
+  url = {https://www.sciencedirect.com/science/article/pii/S0045782522000895},
+  year = {2022},
+  volume = {393},
+  pages = {114726},
+  issn = {0045-7825},
+  journal = {Computer Methods in Applied Mechanics and Engineering},
+}
+```
 
 ----
 
