### 2023-07-31 91f1610
Updated docs CRATE version and JOSS citation.
